requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
aiosmtplib
